requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
scikit-image>=0.18.0
scikit-learn>=1.0.0
Pillow>=8.3.0
tqdm>=4.62.0
PyWavelets>=1.1.0
diffusers>=0.11.0
transformers>=4.15.0
accelerate>=0.12.0
ftfy
regex
requests>=2.25.0
pathlib2>=2.3.0
clip @ git+https://github.com/openai/CLIP.git 